feat: move generating analyzed audit to the resolvers package
